update inventory, room and stay services + migration

+++ app/schemas/stay.py
@@ -1,11 +1,12 @@
 from datetime import datetime
+from decimal import Decimal
 from pydantic import BaseModel, Field, ConfigDict
 
 class StayCreate(BaseModel):
     room_id: int = Field(..., gt=0)
     client_id: int = Field(..., gt=0)
     reservation_id: int | None = Field(default=None, gt=0)
-    price_per_night: float = Field(..., gt=0)
+    price_per_night: Decimal = Field(..., gt=0, decimal_places=2)
     notes: str | None = Field(default=None, max_length=5000)
     checkin_datetime: datetime | None = Field(default=None)
 
@@ -19,7 +20,7 @@ class StayRead(BaseModel):
     created_by: int
     checkin_datetime: datetime
     checkout_datetime: datetime | None
-    price_per_night: float
+    price_per_night: Decimal
     status: str
     notes: str | None
